index.html: indicate that use of Lanyrd is optional
RMS wants folks to be able to register for SeaGL anonymously and not
have to use nonfree software to register.

I just made registration optional for now. Perhaps we'll eventually
require registration, but if we do so we'll do only require the use of
Free software.

I'm not sure what "anonymous registration" entails, but I imagine we can
support that too.