Add note about participating to footer
This adds a note to the footer about how folks who are interested in
helping organize and run SeaGL can reach us and help. We want everyone
who is interested to let us know and also want to ask for specific
skills we're looking for, so I made the call both broad and listed the
roles we need filled. This footer will be included in every SeaGL
website page.