Update volunteer blog post
This updates the volunteers blog post to account for the new tech stack
and speaker-flow. Because we won't have participants on video a video
help-desk is no longer necessary. We will also have SeaGL staff managing
the technical side of the videos and IRC channels, so a per-talk IRC
channel isn't necessary. This updates the blog post to remove the
unnecessary positions, and updates descriptions.