Create Code of Practice Page
This year, we are creating a static Code of Practice page on the site
rather than a blog post. The Code of Practice doesn't really change from
year to year. Instead, we re-affirm our commitment to abide by it. By
having it as a static page, it will be easier to re-use, link to, and
reference.