Document GitHub Pages–specific configuration
GitHub Pages worked around a problem with their platform's handling of
time zones by altering the default behavior of Jekyll:

  - https://github.com/github/pages-gem/pull/302

This created an undocumented divergence between development and
production builds, in which the latter unexpectedly publishes content
that is intended to be hidden.

To restore parity between environments, Jekyll's `future` option must be
explicitly set.